Fix and move renderer tests
