--- firmware/tests/render-balloon/manifest.json
@@ -1,58 +1,0 @@
-{
-    "defines": {
-        "XS_MODS": 1
-    },
-    "include": [
-        "$(MODDABLE)/examples/manifest_base.json",
-        "$(MODDABLE)/examples/manifest_typings.json",
-        "$(MODDABLE)/examples/manifest_commodetto.json",
-        "$(MODDABLE)/modules/commodetto/outline/manifest.json",
-        "$(MODDABLE)/modules/base/deepEqual/manifest.json",
-        "$(MODDABLE)/modules/base/structuredClone/manifest.json",
-        "$(MODULES)/files/preference/manifest.json"
-    ],
-    "modules": {
-        "*": [
-            "./main",
-            "../../stackchan/renderers/*",
-            "../../stackchan/utilities/*"
-        ]
-    },
-    "resources": {
-        "*-mask": [
-            {
-                "source": "$(MODDABLE)/examples/assets/scalablefonts/NotoSans/NotoSansJP-Regular",
-                "size": 24,
-                "blocks": [
-                    "Hiragana",
-                    "Katakana",
-                    "Basic Latin"
-                ],
-                "characters": "えっ今からでも入れる保険があるんですか！？"
-            }
-        ]
-    },
-    "preload": [
-        "consts",
-        "stackchan-util",
-        "face-renderer"
-    ],
-    "config": {
-        "rotation": 90
-    },
-    "creation": {
-        "static": 98304,
-        "chunk": {
-            "initial": 1536,
-            "incremental": 512
-        },
-        "heap": {
-            "initial": 512,
-            "incremental": 64
-        },
-        "stack": 512,
-        "keys": {
-            "available": 512
-        }
-    }
-}
