--- firmware/tests/render-face/manifest.json
@@ -1,29 +1,0 @@
-{
-    "defines": {
-        "XS_MODS": 1
-    },
-    "include": [
-        "$(MODDABLE)/examples/manifest_base.json",
-        "$(MODDABLE)/examples/manifest_typings.json",
-		"$(MODDABLE)/examples/manifest_commodetto.json",
-        "$(MODDABLE)/examples/manifest_piu.json",
-		"$(MODDABLE)/modules/commodetto/outline/manifest.json",
-		"$(MODDABLE)/modules/base/deepEqual/manifest.json",
-		"$(MODDABLE)/modules/base/structuredClone/manifest.json",
-        "$(MODULES)/files/preference/manifest.json"
-    ],
-    "modules": {
-        "*": [
-            "./main",
-            "../../stackchan/renderers/*",
-            "../../stackchan/utilities/*"
-        ]
-    },
-    "preload": [
-		"stackchan-util",
-        "simple-face"
-    ],
-    "config": {
-        "rotation": 90
-    }
-}
