+++ firmware/tests/renderers/render-face/manifest.json
@@ -1,0 +1,15 @@
+{
+    "include": [
+        "$(MODDABLE)/examples/manifest_base.json",
+        "$(MODDABLE)/examples/manifest_typings.json",
+        "../../../stackchan/renderers/manifest_renderer.json"
+    ],
+    "modules": {
+        "*": [
+            "./main"
+        ]
+    },
+    "config": {
+        "rotation": 90
+    }
+}
